+++ src/app/sales/layout/layout.component.ts
@@ -2,6 +2,10 @@ import { MediaMatcher } from '@angular/cdk/layout';
 import { ChangeDetectorRef, Component, EventEmitter, OnDestroy, Output } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { SharedModule } from '../../shared/shared.module';
+import { AuthService } from '../../services/auth.service';
+import { ToastService } from '../../services/toast.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../../confirmation-dialog/confirmation-dialog.component';
 interface SideNavToggle {
   screenWidth: number;
   collapsed: boolean;
@@ -35,7 +39,10 @@ export class SalesLayoutComponent implements OnDestroy {
   constructor(changeDetectorRef: ChangeDetectorRef,
     media: MediaMatcher,
     private activatedRoute: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private authService: AuthService,
+    private toast: ToastService,
+    private dialog: MatDialog,
   ) {
     this.mobileQuery = media.matchMedia('(max-width: 600px)');
     this._mobileQueryListener = () => changeDetectorRef.detectChanges();
@@ -69,7 +76,31 @@ export class SalesLayoutComponent implements OnDestroy {
   }
 
   logout() {
-    sessionStorage.clear();
-    this.router.navigateByUrl('/');
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        title: 'Confirm Logout',
+        message: `Are you sure you want to Logout?`
+      },
+      panelClass: 'custom-dialog-container'
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        let customerId = sessionStorage.getItem('customerId');
+        debugger
+        sessionStorage.clear();
+        this.authService.customerLogout(customerId).subscribe((res: any) => {
+          if (res.message) {
+            sessionStorage.clear();
+            this.router.navigate(['']);
+            this.toast.showToast('warning', "Session Expired Logout Successfully.", "")
+
+          }
+        })
+      }
+    })
+
   }
+
+
 }
